+++ xaeian/serial/__init__.py
@@ -12,10 +12,10 @@ __extras__ = ("serial", ["pyserial"])
 
 from .port import SerialPort, serial_scan
 from .rec import Recorder, MultiRecorder
-from .sh import Shell, convert_value
+from .sh import Shell, convert_value, ihex_to_bin
 
 __all__ = [
   "SerialPort", "serial_scan",
   "Recorder", "MultiRecorder",
-  "Shell", "convert_value",
+  "Shell", "convert_value", "ihex_to_bin",
 ]